@@ -5,13 +5,15 @@
   "dependencies": {
     "@fluent/bundle": "0.17.1",
     "@fluent/langneg": "0.7.0",
-    "@fluent/react": "0.14.1",
+    "@fluent/react": "0.15.2",
     "@sentry/react": "6.19.7",
     "@sentry/tracing": "6.19.7",
     "@yr/monotone-cubic-spline": "1.0.3",
     "audio-recorder-polyfill": "0.4.1",
     "classnames": "2.3.2",
+    "date-fns": "^2.30.0",
     "downshift": "3.4.8",
+    "filesize": "^10.0.7",
     "focus-visible": "5.2.0",
     "intersection-observer": "0.12.0",
     "intl-pluralrules": "1.3.1",
@@ -22,35 +24,38 @@
     "react": "17.0.2",
     "react-balance-text": "2.0.1",
     "react-confetti": "^6.1.0",
-    "react-content-loader": "6.2.0",
+    "react-content-loader": "6.2.1",
     "react-dom": "17.0.2",
+    "react-dropzone": "^14.2.3",
     "react-intersection-observer": "8.33.1",
     "react-modal": "3.14.4",
     "react-redux": "7.2.9",
     "react-router": "5.2.1",
     "react-router-dom": "5.3.0",
     "react-router-navigation-prompt": "1.9.6",
+    "react-tabs": "4.2.1",
     "react-tippy": "1.4.0",
     "redux": "4.1.2",
     "redux-thunk": "2.4.2"
   },
   "devDependencies": {
-    "@babel/core": "7.20.12",
+    "@babel/core": "7.21.3",
     "@babel/preset-env": "7.20.2",
-    "@peculiar/webcrypto": "1.4.1",
+    "@faker-js/faker": "8.3.1",
+    "@peculiar/webcrypto": "1.4.5",
     "@testing-library/react": "12.1.5",
     "@testing-library/react-hooks": "7.0.2",
     "@testing-library/user-event": "13.5.0",
     "@types/classnames": "2.3.1",
-    "@types/jest": "27.4.1",
-    "@types/jest-axe": "3.5.5",
-    "@types/lodash.memoize": "4.1.7",
-    "@types/lodash.pick": "4.4.7",
-    "@types/lodash.throttle": "4.1.7",
+    "@types/jest": "27.5.2",
+    "@types/jest-axe": "3.5.9",
+    "@types/lodash.memoize": "4.1.9",
+    "@types/lodash.pick": "4.4.9",
+    "@types/lodash.throttle": "4.1.9",
     "@types/react": "17.0.53",
     "@types/react-dom": "17.0.14",
     "@types/react-modal": "3.13.1",
-    "@types/react-router": "5.1.18",
+    "@types/react-router": "5.1.20",
     "@types/react-router-dom": "5.3.3",
     "@types/webpack": "5.28.0",
     "@types/webpack-merge": "5.0.0",
@@ -62,7 +67,8 @@
     "copy-webpack-plugin": "10.2.4",
     "css-loader": "6.7.3",
     "cssnano": "5.1.15",
-    "dotenv": "16.0.3",
+    "cypress": "13.5.1",
+    "cypress-file-upload": "5.0.8",
     "file-loader": "6.2.0",
     "html-webpack-plugin": "5.5.0",
     "identity-obj-proxy": "3.0.0",
@@ -73,27 +79,29 @@
     "jest-svg-transformer": "1.0.0",
     "jest-transform-stub": "2.0.0",
     "mini-css-extract-plugin": "2.6.0",
-    "postcss": "8.4.21",
+    "postcss": "8.4.31",
     "postcss-color-mod-function": "3.0.3",
     "postcss-custom-media": "7.0.8",
     "postcss-import": "14.1.0",
     "postcss-loader": "6.2.1",
     "postcss-nested": "5.0.6",
     "postcss-preset-env": "7.4.3",
-    "style-loader": "3.3.1",
+    "style-loader": "3.3.2",
     "stylelint": "14.6.0",
     "stylelint-config-prettier": "9.0.5",
     "stylelint-config-standard": "25.0.0",
     "ts-jest": "27.1.5",
     "ts-loader": "9.2.8",
     "typescript-plugin-css-modules": "3.4.0",
-    "webpack": "5.72.1",
+    "webpack": "5.76.3",
     "webpack-bundle-analyzer": "4.5.0",
     "webpack-cli": "4.9.2"
   },
   "scripts": {
     "build": "webpack --mode=production",
     "build:audit": "AUDIT=true webpack --mode=production",
+    "cypress:open": "cypress open",
+    "cy:run": "cypress run",
     "dev": "webpack --mode=development --watch",
     "lint": "yarn lint:js && yarn lint:css",
     "lint:js": "eslint .",
