@@ -3,7 +3,7 @@ module.exports = {
   env: {
     node: true,
     browser: true,
-    es2021: true,
+    es2022: true
   },
   plugins: ['import', '@typescript-eslint', 'react', 'react-hooks', 'jsx-a11y'],
   extends: [
@@ -18,7 +18,7 @@ module.exports = {
     ecmaFeatures: {
       jsx: true,
     },
-    ecmaVersion: 12,
+    ecmaVersion: 2023,
     sourceType: 'module',
     tsconfigRootDir: __dirname,
     project: [
@@ -26,6 +26,8 @@ module.exports = {
       './common/tsconfig.json',
       './web/tsconfig.json',
       './server/tsconfig.json',
+      './web/cypress/tsconfig.json',
+      './bundler/tsconfig.json'
     ],
   },
   rules: {
@@ -40,6 +42,13 @@ module.exports = {
     // Disabling this because it seems there is no way for ESLint to know we use Localized for translating text
     'jsx-a11y/heading-has-content': 'off',
     'jsx-a11y/anchor-has-content': 'off',
+    // Disabling this because captions for our media files don't make sense for our use-case
+    'jsx-a11y/media-has-caption': 'off',
+    // Disabling unused-vars warning when we destructure an object to remove properties e.g.:
+    // const almostPerfect = {'a': 'awesome', 'b': 'meh', 'c': 'need it'}
+    // const { b, ...perfect } = almostPerfect -- perfect = {'a': 'awesome', 'c': 'need it'}
+    // here we don't want the 'b' property so we separate it out and bundle the rest in 'perfect'
+    '@typescript-eslint/no-unused-vars': ['error', { 'ignoreRestSiblings': true, "varsIgnorePattern": "^_$" }]
   },
   settings: {
     react: {
@@ -52,4 +61,4 @@ module.exports = {
       typescript: {},
     },
   },
-};
+}
