@@ -1,10 +1,10 @@
-export type TaxonomyToken = 'singlewordBenchmark' | 'covid19Spotter';
+export type TaxonomyToken = 'singlewordBenchmark' | 'covid19Spotter'
 
 export type TaxonomyType = {
-  name: string;
-  source: string;
-  locales: string[];
-};
+  name: string
+  source: string
+  locales: string[]
+}
 
 export const taxonomies: { [key in TaxonomyToken]: TaxonomyType } = {
   singlewordBenchmark: {
@@ -30,4 +30,21 @@ export const taxonomies: { [key in TaxonomyToken]: TaxonomyType } = {
     source: 'du-covid-keywords',
     locales: ['rw'],
   },
-};
+}
+
+export const sentenceDomains = [
+  'agriculture',
+  'automotive',
+  'finance',
+  'food_service_retail',
+  'general',
+  'healthcare',
+  'history_law_government',
+  'language_fundamentals',
+  'media_entertainment',
+  'nature_environment',
+  'news_current_affairs',
+  'technology_robotics',
+] as const
+
+export type SentenceDomain = typeof sentenceDomains[number]
