@@ -1,5 +1,4 @@
 import * as fs from 'fs';
-import { S3, SSM } from 'aws-sdk';
 import { SESClientConfig } from '@aws-sdk/client-ses';
 import { config } from 'dotenv';
 
@@ -26,17 +25,14 @@ export type CommonVoiceConfig = {
   MYSQLREPLICAPORT?: number;
   CLIP_BUCKET_NAME: string;
   DATASET_BUCKET_NAME: string;
+  BULK_SUBMISSION_BUCKET_NAME: string;
   AWS_REGION: string;
   ENVIRONMENT: string;
   RELEASE_VERSION?: string;
   SECRET: string;
   AWS_SES_CONFIG: SESClientConfig;
-  S3_CONFIG: S3.Types.ClientConfiguration;
-  S3_LOCAL_DEVELOPMENT_ENDPOINT?: string;
-  CINCHY_CONFIG: S3.Types.ClientConfiguration;
-  CINCHY_ENABLED: boolean;
-  SSM_ENABLED: boolean;
-  SSM_CONFIG: SSM.Types.ClientConfiguration;
+  STORAGE_LOCAL_DEVELOPMENT_ENDPOINT: string;
+  GCP_CREDENTIALS: object;
   ADMIN_EMAILS: string;
   AUTH0: {
     DOMAIN: string;
@@ -83,20 +79,20 @@ const BASE_CONFIG: CommonVoiceConfig = {
     'CV_DATASET_BUCKET_NAME',
     'common-voice-datasets'
   ),
+  BULK_SUBMISSION_BUCKET_NAME: configEntry(
+    'CV_BULK_SUBMISSION_BUCKET_NAME',
+    'common-voice-bulk-submissions'
+  ),
   ENVIRONMENT: configEntry('CV_ENVIRONMENT', 'prod'),
   SECRET: configEntry('CV_SECRET', 'super-secure-secret'),
   ADMIN_EMAILS: configEntry('CV_ADMIN_EMAILS', null),
   AWS_REGION: configEntry('CV_AWS_REGION', 'us-west-2'),
   AWS_SES_CONFIG: configEntry('CV_AWS_SES_CONFIG', {}, castJson),
-  S3_CONFIG: configEntry('CV_S3_CONFIG', {}, castJson),
-  S3_LOCAL_DEVELOPMENT_ENDPOINT: configEntry(
-    'CV_S3_LOCAL_DEVELOPMENT_ENDPOINT',
-    null
+  STORAGE_LOCAL_DEVELOPMENT_ENDPOINT: configEntry(
+    'CV_STORAGE_LOCAL_DEVELOPMENT_ENDPOINT',
+    'http://localhost:8080'
   ),
-  CINCHY_CONFIG: configEntry('CV_CINCHY_CONFIG', {}, castJson),
-  CINCHY_ENABLED: configEntry('CV_CINCHY_ENABLED', false, castBoolean),
-  SSM_ENABLED: configEntry('CV_SSM_ENABLED', false, castBoolean),
-  SSM_CONFIG: configEntry('CV_SSM_CONFIG', {}, castJson),
+  GCP_CREDENTIALS: configEntry('CV_GCP_CREDENTIALS', {}, castJson),
   AUTH0: {
     DOMAIN: configEntry('CV_AUTH0_DOMAIN', ''),
     CLIENT_ID: configEntry('CV_AUTH0_CLIENT_ID', ''),
@@ -121,47 +117,6 @@ const BASE_CONFIG: CommonVoiceConfig = {
 let injectedConfig: CommonVoiceConfig;
 let loadedConfig: CommonVoiceConfig;
 
-const ssm = new SSM(BASE_CONFIG.SSM_CONFIG);
-
-async function getSecret(key: string) {
-  const path = `/voice/${BASE_CONFIG.ENVIRONMENT}/${key}`;
-  const params = {
-    Name: path,
-    WithDecryption: true,
-  };
-  const secret = await ssm.getParameter(params).promise();
-
-  return secret.Parameter.Value;
-}
-
-let loadedSecrets: Partial<CommonVoiceConfig>;
-
-export async function getSecrets(): Promise<Partial<CommonVoiceConfig>> {
-  if (loadedSecrets) {
-    console.log('Use pre-loaded secrets');
-    return loadedSecrets;
-  }
-
-  loadedSecrets = {};
-
-  if (BASE_CONFIG.SSM_ENABLED) {
-    console.log('Fetch SSM secrets.');
-    loadedSecrets = {
-      MYSQLPASS: await getSecret('mysql-user-pw'),
-      DB_ROOT_PASS: await getSecret('mysql-root-pw'),
-      MYSQLHOST: await getSecret('mysql-host'),
-      SECRET: await getSecret('app-secret'),
-      BASKET_API_KEY: await getSecret('basket-api-key'),
-      AUTH0: {
-        DOMAIN: await getSecret('auth0-domain'),
-        CLIENT_ID: await getSecret('auth0-client-id'),
-        CLIENT_SECRET: await getSecret('auth0-client-secret'),
-      },
-    };
-  }
-  return loadedSecrets;
-}
-
 export function injectConfig(config: Partial<CommonVoiceConfig>) {
   injectedConfig = { ...BASE_CONFIG, ...config };
 }
@@ -185,7 +140,7 @@ export function getConfig(): CommonVoiceConfig {
       `Could not load config.json, using defaults (error message: ${err.message})`
     );
   }
-  loadedConfig = { ...BASE_CONFIG, ...loadedSecrets, ...fileConfig };
+  loadedConfig = { ...BASE_CONFIG, ...fileConfig };
 
   return loadedConfig;
 }
