@@ -39,35 +39,41 @@ const ListenPage = React.lazy(
   () => import('./pages/contribution/listen/listen')
 );
 const SpeakPage = React.lazy(() => import('./pages/contribution/speak/speak'));
-const DemoPage = React.lazy(() => import('./layout/demo-layout'));
+const WritePage = React.lazy(
+  () => import('./pages/contribution/sentence-collector/write/write-container')
+)
+const ReviewPage = React.lazy(
+  () => import('./pages/contribution/sentence-collector/review/review')
+)
+const DemoPage = React.lazy(() => import('./layout/demo-layout'))
 
-const SentryRoute = Sentry.withSentryRouting(Route);
+const SentryRoute = Sentry.withSentryRouting(Route)
 
 const SENTRY_DSN_WEB =
-  'https://40742891598c4900aacac78dd1145d7e@o1069899.ingest.sentry.io/6251028';
+  'https://40742891598c4900aacac78dd1145d7e@o1069899.ingest.sentry.io/6251028'
 
 Sentry.init({
   dsn: shouldEmitErrors() ? SENTRY_DSN_WEB : null,
   integrations: [new BrowserTracing()],
   environment: isProduction() ? 'prod' : 'stage',
   release: process.env.GIT_COMMIT_SHA || null,
-});
+})
 
 interface PropsFromState {
-  api: API;
-  account: UserClient;
-  notifications: Notifications.State;
-  uploads: Uploads.State;
-  languages: Languages.State;
-  messageOverwrites: Flags.MessageOverwrites;
+  api: API
+  account: UserClient
+  notifications: Notifications.State
+  uploads: Uploads.State
+  languages: Languages.State
+  messageOverwrites: Flags.MessageOverwrites
 }
 
 interface PropsFromDispatch {
-  addNotification: typeof Notifications.actions.addBanner;
-  removeUpload: typeof Uploads.actions.remove;
-  setLocale: typeof Locale.actions.set;
-  refreshUser: typeof User.actions.refresh;
-  updateUser: typeof User.actions.update;
+  addNotification: typeof Notifications.actions.addBanner
+  removeUpload: typeof Uploads.actions.remove
+  setLocale: typeof Locale.actions.set
+  refreshUser: typeof User.actions.refresh
+  updateUser: typeof User.actions.update
 }
 
 interface LocalizedPagesProps
@@ -78,51 +84,51 @@ interface LocalizedPagesProps
     RouteComponentProps<any, any, any> {}
 
 interface LocalizedPagesState {
-  uploadPercentage?: number;
+  uploadPercentage?: number
 }
 
 let LocalizedPage: any = class extends React.Component<
   LocalizedPagesProps,
   LocalizedPagesState
 > {
-  seenAwardIds: number[] = [];
+  seenAwardIds: number[] = []
   state: LocalizedPagesState = {
     uploadPercentage: null,
-  };
+  }
 
-  isUploading = false;
+  isUploading = false
 
   async componentDidMount() {
-    this.props.updateUser({});
-    this.props.refreshUser();
+    this.props.updateUser({})
+    this.props.refreshUser()
 
     if (isMobileSafari()) {
-      document.body.classList.add('mobile-safari');
+      document.body.classList.add('mobile-safari')
     }
 
-    Modal.setAppElement('#root');
+    Modal.setAppElement('#root')
   }
 
   async UNSAFE_componentWillReceiveProps(nextProps: LocalizedPagesProps) {
-    const { account, addNotification, api, uploads } = nextProps;
+    const { account, addNotification, api, uploads } = nextProps
 
-    this.runUploads(uploads).catch(e => console.error(e));
+    this.runUploads(uploads).catch(e => console.error(e))
 
     window.onbeforeunload =
       uploads.length > 0
         ? (e: any) =>
             (e.returnValue =
               'Leaving the page now aborts pending uploads. Are you sure?')
-        : undefined;
+        : undefined
 
     const award = account?.awards
       ? account.awards.find(
           a => !a.notification_seen_at && !this.seenAwardIds.includes(a.id)
         )
-      : null;
+      : null
 
     if (award) {
-      this.seenAwardIds.push(...account.awards.map(a => a.id));
+      this.seenAwardIds.push(...account.awards.map(a => a.id))
       addNotification(
         `Success, ${award.amount} Clip ${
           award.days_interval == 1 ? 'daily' : 'weekly'
@@ -135,39 +141,55 @@ let LocalizedPage: any = class extends React.Component<
             },
           ],
         }
-      );
-      await api.seenAwards('notification');
+      )
+      await api.seenAwards('notification')
     }
   }
 
   async runUploads(uploads: Uploads.State) {
-    if (this.isUploading) return;
-    this.isUploading = true;
-    this.setState({ uploadPercentage: 0 });
+    if (this.isUploading) return
+    this.isUploading = true
+    this.setState({ uploadPercentage: 0 })
     for (let i = 0; i < uploads.length; i++) {
-      this.setState({ uploadPercentage: (i + 1) / (uploads.length + 1) });
-      const upload = uploads[i];
+      this.setState({ uploadPercentage: (i + 1) / (uploads.length + 1) })
+      const upload = uploads[i]
       try {
-        await upload();
+        await upload()
       } catch (e) {
-        console.error('upload error', e);
+        console.error('upload error', e)
       }
-      this.props.removeUpload(upload);
+      this.props.removeUpload(upload)
     }
-    this.setState({ uploadPercentage: null });
-    this.isUploading = false;
+    this.setState({ uploadPercentage: null })
+    this.isUploading = false
 
     if (this.props.uploads.length > 0) {
-      await this.runUploads(this.props.uploads);
+      await this.runUploads(this.props.uploads)
     }
+  }
+
+  shouldRedirectToRoot({
+    isContributable,
+    route,
+  }: {
+    isContributable: boolean
+    route: string
+  }) {
+    if (isContributable && (route === URLS.SPEAK || route === URLS.LISTEN)) {
+      return false
+    } else if (route === URLS.WRITE || route === URLS.REVIEW) {
+      return false
+    }
+
+    return true
   }
 
   render() {
     const { locale, notifications, toLocaleRoute, location, languages } =
-      this.props;
-    const { uploadPercentage } = this.state;
+      this.props
+    const { uploadPercentage } = this.state
 
-    const isContributable = languages.contributableLocales.includes(locale);
+    const isContributable = languages.contributableLocales.includes(locale)
 
     return (
       <>
@@ -208,18 +230,20 @@ let LocalizedPage: any = class extends React.Component<
           {[
             { route: URLS.SPEAK, Component: SpeakPage },
             { route: URLS.LISTEN, Component: ListenPage },
+            { route: URLS.WRITE, Component: WritePage },
+            { route: URLS.REVIEW, Component: ReviewPage },
           ].map(({ route, Component }: any) => (
             <SentryRoute
               key={route}
               exact
               path={toLocaleRoute(route)}
               render={props =>
-                isContributable ? (
+                this.shouldRedirectToRoot({ isContributable, route }) ? (
+                  <Redirect to={toLocaleRoute(URLS.ROOT)} />
+                ) : (
                   <Layout shouldHideFooter>
                     <Component {...props} />
                   </Layout>
-                ) : (
-                  <Redirect to={toLocaleRoute(URLS.ROOT)} />
                 )
               }
             />
@@ -227,9 +251,9 @@ let LocalizedPage: any = class extends React.Component<
           {location.pathname.includes(URLS.DEMO) ? <DemoPage /> : <Layout />}
         </Switch>
       </>
-    );
+    )
   }
-};
+}
 
 LocalizedPage.displayName = 'LocalizedPage';
 
