@@ -21,13 +21,19 @@ const ProfileLayoutPage = React.lazy(() => import('../pages/profile/layout'));
 const AboutPage = React.lazy(() => import('../pages/about/about'));
 const LandingPage = React.lazy(() => import('../pages/landing/landing'));
 const ErrorPage = React.lazy(() => import('../pages/error-page/error-page'));
-const CriteriaPage = React.lazy(() => import('../pages/criteria/criteria'));
 const PartnerPage = React.lazy(() => import('../pages/partner/partner'));
+const GuidelinesPage = React.lazy(
+  () => import('../pages/guidelines/guidelines')
+);
+const SentenceCollectorRedirectPage = React.lazy(
+  () =>
+    import('../pages/sentence-collector-redirect/sentence-collector-redirect')
+)
 
-const SentryRoute = Sentry.withSentryRouting(Route);
+const SentryRoute = Sentry.withSentryRouting(Route)
 
 export default function Content({ location }: { location: any }) {
-  const [locale, toLocaleRoute] = useLocale();
+  const [locale, toLocaleRoute] = useLocale()
   return (
     <React.Suspense fallback={<Spinner />}>
       <Switch>
@@ -102,7 +108,7 @@ export default function Content({ location }: { location: any }) {
         />
         <SentryRoute
           path={toLocaleRoute(URLS.CRITERIA)}
-          component={CriteriaPage}
+          render={() => <Redirect to={toLocaleRoute(URLS.GUIDELINES)} />}
         />
         {[URLS.CHALLENGE, URLS.STATS, URLS.GOALS, URLS.AWARDS].map(path => (
           <SentryRoute
@@ -167,24 +173,43 @@ export default function Content({ location }: { location: any }) {
           exact
           path={toLocaleRoute(URLS.SPEAK)}
           render={() => {
-            // note: this is redundant with routing in LocalizedPage in app.tsx, and hanldes
+            // note: this is redundant with routing in LocalizedPage in app.tsx, and handles
             // locale changing edge cases where toLocaleRoute is still using the old locale
-            return <Redirect to={toLocaleRoute(URLS.SPEAK)} />;
+            return <Redirect to={toLocaleRoute(URLS.SPEAK)} />
           }}
         />
         <SentryRoute
           exact
           path={toLocaleRoute(URLS.LISTEN)}
           render={() => {
-            // note: this is redundant with routing in LocalizedPage in app.tsx, and hanldes
+            // note: this is redundant with routing in LocalizedPage in app.tsx, and handles
             // locale changing edge cases where toLocaleRoute is still using the old locale
-            return <Redirect to={toLocaleRoute(URLS.LISTEN)} />;
+            return <Redirect to={toLocaleRoute(URLS.LISTEN)} />
           }}
         />
         <SentryRoute
           exact
           path={toLocaleRoute(URLS.PARTNER)}
           component={PartnerPage}
+        />
+        <SentryRoute
+          exact
+          path={toLocaleRoute(URLS.WRITE)}
+          render={() => {
+            // note: this is redundant with routing in LocalizedPage in app.tsx, and handles
+            // locale changing edge cases where toLocaleRoute is still using the old locale
+            return <Redirect to={toLocaleRoute(URLS.WRITE)} />;
+          }}
+        />
+        <SentryRoute
+          exact
+          path={toLocaleRoute(URLS.GUIDELINES)}
+          component={GuidelinesPage}
+        />
+        <SentryRoute
+          exact
+          path={toLocaleRoute(URLS.SENTENCE_COLLECTOR_REDIRECT)}
+          component={SentenceCollectorRedirectPage}
         />
         <Route
           render={() => (
@@ -193,5 +218,5 @@ export default function Content({ location }: { location: any }) {
         />
       </Switch>
     </React.Suspense>
-  );
+  )
 }
