@@ -1,8 +1,8 @@
-import * as React from 'react';
-import { Localized } from '@fluent/react';
+import * as React from 'react'
+import { Localized } from '@fluent/react'
 
 const HowAddSentences = React.memo(() => {
-  const strong = <strong />;
+  const strong = <strong />
 
   return (
     <>
@@ -10,11 +10,18 @@ const HowAddSentences = React.memo(() => {
         id="about-playbook-how-add-sentences-content-1"
         elems={{
           strong,
-          scAccountLink: (
+          writePageLink: (
             <a
-              href="https://commonvoice.mozilla.org/sentence-collector/login"
+              href="https://commonvoice.mozilla.org/write"
               target="_blank"
-              rel="noopener noreferer"
+              rel="noopener noreferrer"
+            />
+          ),
+          reviewPageLink: (
+            <a
+              href="https://commonvoice.mozilla.org/review"
+              target="_blank"
+              rel="noopener noreferrer"
             />
           ),
         }}>
@@ -27,21 +34,21 @@ const HowAddSentences = React.memo(() => {
             <a
               href="https://commonvoice.mozilla.org/sentence-collector/#/en/how-to"
               target="_blank"
-              rel="noopener noreferer"
+              rel="noopener noreferrer"
             />
           ),
           licenseLink: (
             <a
               href="https://creativecommons.org/share-your-work/public-domain/cc0/"
               target="_blank"
-              rel="noopener noreferer"
+              rel="noopener noreferrer"
             />
           ),
           waiverLink: (
             <a
               href="https://common-voice.github.io/community-playbook/sub_pages/cc0waiver_process.html"
               target="_blank"
-              rel="noopener noreferer"
+              rel="noopener noreferrer"
             />
           ),
         }}>
@@ -55,18 +62,16 @@ const HowAddSentences = React.memo(() => {
             <a
               href="https://discourse.mozilla.org/t/sentence-extractor-current-status-and-workflow-summary/62332"
               target="_blank"
-              rel="noopener noreferer"
+              rel="noopener noreferrer"
             />
           ),
         }}>
         <p />
       </Localized>
-
-      <Localized id="about-playbook-how-add-sentences-content-4">
-        <p />
-      </Localized>
     </>
-  );
-});
+  )
+})
 
-export default HowAddSentences;
+HowAddSentences.displayName = 'HowAddSentences'
+
+export default HowAddSentences
