@@ -5,6 +5,7 @@ import {
 } from '@fluent/react';
 import * as React from 'react';
 import { connect } from 'react-redux';
+import { Tooltip } from 'react-tippy';
 import { Flags } from '../../../stores/flags';
 import { Locale } from '../../../stores/locale';
 import StateTree from '../../../stores/tree';
@@ -23,8 +24,9 @@ import {
   SkipIcon,
   ExternalLinkIcon,
   ArrowLeft,
+  QuestionIcon,
 } from '../../ui/icons';
-import { Button, StyledLink, LabeledCheckbox } from '../../ui/ui';
+import { Button, StyledLink, LabeledCheckbox, LinkButton } from '../../ui/ui';
 import { PrimaryButton } from '../../primary-buttons/primary-buttons';
 import ShareModal from '../../share-modal/share-modal';
 import { ReportButton, ReportModal, ReportModalProps } from './report/report';
@@ -382,7 +384,12 @@ class ContributionPage extends React.Component<ContributionPageProps, State> {
             <div className="cards-and-instruction">
               {instruction({
                 vars: { actionType: getString('action-click') },
-                children: <div className="instruction hidden-sm-down" />,
+                children: (
+                  <div
+                    className="instruction hidden-sm-down"
+                    data-testid="instruction"
+                  />
+                ),
               }) || <div className="instruction hidden-sm-down" />}
 
               <div className="cards">
@@ -410,7 +417,8 @@ class ContributionPage extends React.Component<ContributionPageProps, State> {
                           }%)`,
                         ].join(' '),
                         opacity: i < activeSentenceIndex ? 0 : 1,
-                      }}>
+                      }}
+                      data-testid={`card-${i + 1}`}>
                       <div style={{ margin: 'auto', width: '100%' }}>
                         {sentence?.text}
                         {sentence?.taxonomy ? (
@@ -498,7 +506,7 @@ class ContributionPage extends React.Component<ContributionPageProps, State> {
         {!hasErrors && !isSubmitted && (
           <LocaleLink
             blank
-            to={URLS.CRITERIA}
+            to={URLS.GUIDELINES}
             className="contribution-criteria hidden-md-up">
             <ExternalLinkIcon />
             <Localized id="contribution-criteria-link" />
@@ -507,20 +515,30 @@ class ContributionPage extends React.Component<ContributionPageProps, State> {
 
         <div className="buttons">
           <div>
-            <Button
+            <LinkButton
               rounded
               outline
-              className="hidden-sm-down"
-              onClick={this.toggleShortcutsModal}>
-              <KeyboardIcon />
-              <Localized id="shortcuts">
+              className="guidelines-button"
+              blank
+              to={URLS.GUIDELINES}>
+              <QuestionIcon />
+              <Localized id="guidelines">
                 <span />
               </Localized>
-            </Button>
-            <div className="extra-button">
+            </LinkButton>
+            <div className="extra-buttons">
               <ReportButton
                 onClick={() => this.setState({ showReportModal: true })}
               />
+              <Tooltip title="Shortcuts" arrow>
+                <Button
+                  rounded
+                  outline
+                  className="hidden-md-down shortcuts-btn"
+                  onClick={this.toggleShortcutsModal}>
+                  <KeyboardIcon />
+                </Button>
+              </Tooltip>
             </div>
           </div>
           <div>
@@ -533,11 +551,12 @@ class ContributionPage extends React.Component<ContributionPageProps, State> {
                 'fs-ignore-rage-clicks',
               ].join(' ')}
               disabled={!this.isLoaded}
-              onClick={onSkip}>
+              onClick={onSkip}
+              data-testid="skip-button">
+              <SkipIcon />
               <Localized id="skip">
                 <span />
               </Localized>{' '}
-              <SkipIcon />
             </Button>
             {onSubmit && shouldHideCTA && (
               <form
@@ -570,6 +589,7 @@ class ContributionPage extends React.Component<ContributionPageProps, State> {
                     ].join(' ')}
                     disabled={!this.isDone}
                     type="submit"
+                    data-testid="submit-button"
                   />
                 </Localized>
               </form>
