@@ -7,7 +7,6 @@ import * as React from 'react';
 import { useEffect, useRef, useState } from 'react';
 import { DAILY_GOALS } from '../../../constants';
 import { useAccount, useAPI } from '../../../hooks/store-hooks';
-import { useTypedSelector } from '../../../stores/tree';
 import URLS from '../../../urls';
 import { LocaleLink, useLocale } from '../../locale-helpers';
 import { CheckIcon, MicIcon, PlayOutlineIcon } from '../../ui/icons';
@@ -118,7 +117,7 @@ function Success({
   );
 
   return (
-    <div className="contribution-success">
+    <div className="contribution-success" data-testid="contribution-success">
       <div className="counter done">
         <CheckIcon />
         <Localized
@@ -177,7 +176,7 @@ function Success({
         </div>
       )}
 
-      <ContributeMoreButton>
+      <ContributeMoreButton data-testid="contribute-more-button">
         {type === 'speak' ? <MicIcon /> : <PlayOutlineIcon />}
         <Localized id="contribute-more" vars={{ count: SET_COUNT }}>
           <span />
