@@ -10,26 +10,24 @@ import useSortedLocales from '../../../hooks/use-sorted-locales';
 import { LabeledSelect, Spinner } from '../../ui/ui';
 
 import DatasetDownloadEmailPrompt from './dataset-download-email-prompt';
-
-import './dataset-corpus-download.css';
 import { useAPI } from '../../../hooks/store-hooks';
 import DatasetCorpusDownloadTable from './dataset-corpus-download-table';
 import PageHeading from '../../ui/page-heading';
 import { formatBytes } from '../../../utility';
 import { DeltaReadMoreLink } from '../../shared/links';
+import { MobileDatasetMetadataViewer } from './metadata-viewer/mobile/metadata-viewer';
+
+import { LanguageDataset } from './metadata-viewer/types';
+
+import { DesktopMetaDataViewer } from './metadata-viewer/desktop/metadata-viewer';
+
+import './dataset-corpus-download.css';
 
 interface Props extends WithLocalizationProps {
   languagesWithDatasets: { id: number; name: string }[];
   initialLanguage: string;
   isSubscribedToMailingList: boolean;
 }
-
-type LanguageDatasets = {
-  download_path: string;
-  id: number;
-  checksum: string;
-  size: number;
-};
 
 const DatasetCorpusDownload = ({
   getString,
@@ -38,10 +36,12 @@ const DatasetCorpusDownload = ({
   isSubscribedToMailingList,
 }: Props) => {
   const [isLoading, setIsLoading] = useState(true);
-  const [selectedDataset, setSelectedDataset] = useState<LanguageDatasets>();
-  const [languageDatasets, setLanguageDatasets] = useState<LanguageDatasets[]>(
+  const [selectedDataset, setSelectedDataset] = useState<LanguageDataset>();
+  const [languageDatasets, setLanguageDatasets] = useState<LanguageDataset[]>(
     []
   );
+
+  const [selectedTableRowIndex, setSelectedTableRowIndex] = useState(0);
   const api = useAPI();
 
   const [locale, setLocale] = useState(initialLanguage);
@@ -55,15 +55,17 @@ const DatasetCorpusDownload = ({
     setLocale(newLocale);
   };
 
-  const handleRowSelect = (selectedId: number) =>
+  const handleRowSelect = (selectedId: number, index: number) => {
     setSelectedDataset(languageDatasets.find(d => d.id === selectedId));
+    setSelectedTableRowIndex(index);
+  };
 
   useEffect(() => {
     setIsLoading(true);
 
     api.getLanguageDatasetStats(locale).then(data => {
       setLanguageDatasets(
-        data.filter((dataset: LanguageDatasets) => !!dataset.download_path)
+        data.filter((dataset: LanguageDataset) => !!dataset.download_path)
       );
       setSelectedDataset(data[0]);
       setIsLoading(false);
@@ -101,30 +103,38 @@ const DatasetCorpusDownload = ({
             ))}
           </LabeledSelect>
         </div>
-        <div
-          style={{
-            display: 'flex',
-            width: '100%',
-            alignItems: 'start',
-            flexDirection: 'column',
-          }}>
-          {isLoading && <Spinner />}
-          {!isLoading && languageDatasets && (
-            <DatasetCorpusDownloadTable
-              onRowSelect={handleRowSelect}
-              releaseData={languageDatasets}
-              selectedId={selectedDataset?.id || languageDatasets[0].id}
-            />
-          )}
+        <div className="table-metadata-container">
+          <div className="table-email-prompt-container">
+            {isLoading && <Spinner />}
+            {!isLoading && languageDatasets && (
+              <DatasetCorpusDownloadTable
+                onRowSelect={handleRowSelect}
+                releaseData={languageDatasets}
+                selectedId={selectedDataset?.id || languageDatasets[0].id}
+              />
+            )}
 
-          {selectedDataset && selectedDataset.download_path && (
-            <DatasetDownloadEmailPrompt
-              selectedLocale={locale}
-              downloadPath={selectedDataset.download_path}
-              releaseId={selectedDataset.id.toString()}
-              checksum={selectedDataset.checksum}
-              size={formatBytes(selectedDataset.size, initialLanguage)}
-              isSubscribedToMailingList={isSubscribedToMailingList}
+            {!isLoading && languageDatasets && (
+              <MobileDatasetMetadataViewer releaseData={languageDatasets} />
+            )}
+
+            {selectedDataset && selectedDataset.download_path && (
+              <DatasetDownloadEmailPrompt
+                selectedLocale={locale}
+                downloadPath={selectedDataset.download_path}
+                releaseId={selectedDataset.id.toString()}
+                checksum={selectedDataset.checksum}
+                size={formatBytes(selectedDataset.size, initialLanguage)}
+                isSubscribedToMailingList={isSubscribedToMailingList}
+              />
+            )}
+          </div>
+
+          {!isLoading && selectedDataset && selectedDataset.splits && (
+            <DesktopMetaDataViewer
+              selectedTableRowIndex={selectedTableRowIndex}
+              datasetsCount={languageDatasets.length}
+              splits={selectedDataset.splits}
             />
           )}
         </div>
