@@ -2,11 +2,12 @@ import { Localized } from '@fluent/react';
 import * as React from 'react';
 import { useState } from 'react';
 import { shallowEqual } from 'react-redux';
+import classNames from 'classnames';
+
 import { trackHome } from '../../../services/tracker';
 import { useTypedSelector } from '../../../stores/tree';
 import { ContributableLocaleLock } from '../../locale-helpers';
 import { RecordLink } from '../../primary-buttons/primary-buttons';
-import RegisterSection from '../../register-section/register-section';
 import { LinkButton } from '../../ui/ui';
 import Page from '../../ui/page';
 import Hero from './hero';
@@ -74,7 +75,6 @@ export default function HomePage() {
           )
         }
       />
-
       <div className="text">
         <div className="inner">
           <div className="title">
@@ -121,13 +121,12 @@ export default function HomePage() {
           </div>
         </div>
       </div>
-
-      <div className="stats">
+      <div className={classNames('stats', { 'logged-in': user.account })}>
         <ClipsStats.Root />
         <VoiceStats />
       </div>
 
-      {user.account ? (
+      {user.account && (
         <section className="contribute-section">
           <div className="mars-container">
             <img src="/img/mars.svg" alt="Mars" />
@@ -163,22 +162,6 @@ export default function HomePage() {
             />
           </div>
         </section>
-      ) : (
-        <RegisterSection marsSrc="/img/mars.svg">
-          <h1>
-            <Localized id="help-make-dataset" />
-          </h1>
-          <h2>
-            <Localized id="profile-not-required" />
-          </h2>
-          <Localized id="sign-up-account">
-            <LinkButton
-              rounded
-              href="/login"
-              onClick={() => trackHome('click-benefits-register', locale)}
-            />
-          </Localized>
-        </RegisterSection>
       )}
     </Page>
   );
