@@ -10,42 +10,48 @@ import {
   Sentence,
   Clip,
   UserLanguage,
-} from 'common';
-import { Locale } from '../stores/locale';
-import { User } from '../stores/user';
-import { USER_KEY } from '../stores/root';
+  SentenceSubmission,
+  SentenceVote,
+  TakeoutResponse,
+} from 'common'
+import { Locale } from '../stores/locale'
+import { User } from '../stores/user'
+import { USER_KEY } from '../stores/root'
 
 interface FetchOptions {
-  method?: 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
-  isJSON?: boolean;
+  method?: 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH'
+  isJSON?: boolean
   headers?: {
-    [headerName: string]: string;
-  };
-  body?: any; // eslint-disable-line @typescript-eslint/no-explicit-any
+    [headerName: string]: string
+  }
+  body?: any // eslint-disable-line @typescript-eslint/no-explicit-any
+  signal?: AbortSignal
 }
 
 interface Vote extends Event {
-  hasEarnedSessionToast?: boolean;
-  showFirstContributionToast?: boolean;
-  showFirstStreakToast?: boolean;
-  challengeEnded?: boolean;
+  hasEarnedSessionToast?: boolean
+  showFirstContributionToast?: boolean
+  showFirstStreakToast?: boolean
+  challengeEnded?: boolean
 }
 
-const API_PATH = location.origin + '/api/v1';
+const API_PATH = location.origin + '/api/v1'
 
 const getChallenge = (user: User.State): string => {
   return user?.account?.enrollment?.challenge
     ? user.account.enrollment.challenge
-    : null;
-};
+    : null
+}
 
 export default class API {
-  private readonly locale: Locale.State;
-  private readonly user: User.State;
+  private readonly locale: Locale.State
+  private readonly user: User.State
+  private readonly abortController: AbortController
 
   constructor(locale: Locale.State, user: User.State) {
-    this.locale = locale;
-    this.user = user;
+    this.locale = locale
+    this.user = user
+    this.abortController = new AbortController()
   }
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -55,7 +61,7 @@ export default class API {
         isJSON: true,
       },
       options
-    );
+    )
 
     const finalHeaders = Object.assign(
       isJSON
@@ -64,12 +70,12 @@ export default class API {
           }
         : {},
       headers
-    );
+    )
 
-    const { user } = this;
+    const { user } = this
     if (path.startsWith(location.origin) && !user.account && user.userId) {
       finalHeaders['Authorization'] =
-        'Basic ' + btoa(user.userId + ':' + user.authToken);
+        'Basic ' + btoa(user.userId + ':' + user.authToken)
     }
 
     const response = await fetch(path, {
@@ -81,39 +87,46 @@ export default class API {
           ? body
           : JSON.stringify(body)
         : undefined,
-    });
+    })
+
     if (response.status == 401) {
-      localStorage.removeItem(USER_KEY);
-      location.reload();
-      return;
+      localStorage.removeItem(USER_KEY)
+      location.reload()
+      return
     }
+
+    if (response.status === 429) {
+      throw new Error(response.statusText)
+    }
+
     if (response.status >= 400) {
       if (response.statusText.includes('save_clip_error')) {
-        throw new Error(response.statusText);
+        throw new Error(response.statusText)
       }
-      throw new Error(await response.text());
+      throw new Error(await response.text())
     }
-    return isJSON ? response.json() : response.text();
+
+    return isJSON ? response.json() : response.text()
   }
 
   forLocale(locale: string) {
-    return new API(locale, this.user);
+    return new API(locale, this.user)
   }
 
   getLocalePath() {
-    return this.locale ? API_PATH + '/' + this.locale : API_PATH;
+    return this.locale ? API_PATH + '/' + this.locale : API_PATH
   }
 
   getClipPath() {
-    return this.getLocalePath() + '/clips';
+    return this.getLocalePath() + '/clips'
   }
 
   async fetchRandomSentences(count = 1): Promise<Sentence[]> {
-    return this.fetch(`${this.getLocalePath()}/sentences?count=${count}`);
+    return this.fetch(`${this.getLocalePath()}/sentences?count=${count}`)
   }
 
   async fetchRandomClips(count = 1): Promise<Clip[]> {
-    return this.fetch(`${this.getClipPath()}?count=${count}`);
+    return this.fetch(`${this.getClipPath()}?count=${count}`)
   }
 
   uploadClip(
@@ -121,10 +134,10 @@ export default class API {
     sentenceId: string,
     fromDemo?: boolean
   ): Promise<{
-    showFirstContributionToast?: boolean;
-    hasEarnedSessionToast?: boolean;
-    showFirstStreakToast?: boolean;
-    challengeEnded: boolean;
+    showFirstContributionToast?: boolean
+    hasEarnedSessionToast?: boolean
+    showFirstStreakToast?: boolean
+    challengeEnded: boolean
   }> {
     // make sure nginx server has allow_underscore turned on
     return this.fetch(this.getClipPath(), {
@@ -137,7 +150,7 @@ export default class API {
         source: 'web',
       },
       body: blob,
-    });
+    })
   }
   saveVote(id: string, isValid: boolean): Promise<Vote> {
     return this.fetch(`${this.getClipPath()}/${id}/votes`, {
@@ -146,29 +159,29 @@ export default class API {
         isValid,
         challenge: getChallenge(this.user),
       },
-    });
+    })
   }
 
   fetchDailyClipsCount(): Promise<number> {
-    return this.fetch(this.getClipPath() + '/daily_count');
+    return this.fetch(this.getClipPath() + '/daily_count')
   }
 
   fetchDailyVotesCount(): Promise<number> {
-    return this.fetch(this.getClipPath() + '/votes/daily_count');
+    return this.fetch(this.getClipPath() + '/votes/daily_count')
   }
 
   fetchLocaleMessages(locale: string): Promise<string> {
-    return this.fetch(`/locales/${locale}/messages.ftl`, { isJSON: false });
+    return this.fetch(`/locales/${locale}/messages.ftl`, { isJSON: false })
   }
 
   async fetchCrossLocaleMessages(): Promise<string[][]> {
     return Object.entries(
       await this.fetch(`/cross-locale-messages.json`)
-    ) as string[][];
+    ) as string[][]
   }
 
   fetchRequestedLanguages(): Promise<string[]> {
-    return this.fetch(`${API_PATH}/requested_languages`);
+    return this.fetch(`${API_PATH}/requested_languages`)
   }
 
   requestLanguage(language: string): Promise<void> {
@@ -177,96 +190,94 @@ export default class API {
       body: {
         language,
       },
-    });
+    })
   }
 
   async fetchAllLanguages(): Promise<Language[]> {
-    return this.fetch(`${API_PATH}/languages`);
+    return this.fetch(`${API_PATH}/languages`)
   }
 
   async fetchLanguageStats(): Promise<LanguageStatistics[]> {
-    return this.fetch(`${API_PATH}/stats/languages`);
+    return this.fetch(`${API_PATH}/stats/languages`)
   }
 
   fetchDocument(
     name: 'privacy' | 'terms' | 'challenge-terms'
   ): Promise<string> {
-    const locale = name === 'challenge-terms' ? 'en' : this.locale;
-    return this.fetch(`/${name}/${locale}.html`, { isJSON: false });
+    const locale = name === 'challenge-terms' ? 'en' : this.locale
+    return this.fetch(`/${name}/${locale}.html`, { isJSON: false })
   }
 
   skipSentence(id: string) {
     return this.fetch(`${API_PATH}/skipped_sentences/` + id, {
       method: 'POST',
-    });
+    })
   }
 
   skipClip(id: string) {
     return this.fetch(`${API_PATH}/skipped_clips/` + id, {
       method: 'POST',
-    });
+    })
   }
 
   fetchClipsStats(locale?: string): Promise<
     {
-      date: string;
-      total: number;
-      valid: number;
+      date: string
+      total: number
+      valid: number
     }[]
   > {
-    return this.fetch(API_PATH + (locale ? '/' + locale : '') + '/clips/stats');
+    return this.fetch(API_PATH + (locale ? '/' + locale : '') + '/clips/stats')
   }
 
   fetchClipVoices(locale?: string): Promise<
     {
-      date: string;
-      value: number;
+      date: string
+      value: number
     }[]
   > {
-    return this.fetch(
-      API_PATH + (locale ? '/' + locale : '') + '/clips/voices'
-    );
+    return this.fetch(API_PATH + (locale ? '/' + locale : '') + '/clips/voices')
   }
 
   fetchContributionActivity(
     from: 'you' | 'everyone',
     locale?: string
   ): Promise<{ date: string; value: number }[]> {
-    let endpoint = API_PATH;
+    let endpoint = API_PATH
 
     if (locale) {
-      endpoint += `/${locale}`;
+      endpoint += `/${locale}`
     }
 
-    return this.fetch(`${endpoint}/contribution_activity?from=${from}`);
+    return this.fetch(`${endpoint}/contribution_activity?from=${from}`)
   }
 
   fetchUserClients(): Promise<UserClient[]> {
-    return this.fetch(API_PATH + '/user_clients');
+    return this.fetch(API_PATH + '/user_clients')
   }
 
   fetchAccount(): Promise<UserClient> {
-    return this.fetch(API_PATH + '/user_client');
+    return this.fetch(API_PATH + '/user_client')
   }
 
   saveAccount(data: UserClient): Promise<UserClient> {
     return this.fetch(API_PATH + '/user_client', {
       method: 'PATCH',
       body: data,
-    });
+    })
   }
 
   saveAnonymousAccountLanguages(data: {
-    languages: UserLanguage[];
+    languages: UserLanguage[]
   }): Promise<UserClient> {
     return this.fetch(`${API_PATH}/anonymous_user`, {
       method: 'PATCH',
       body: data,
-    });
+    })
   }
 
   subscribeToNewsletter(email: string): Promise<void> {
-    return this.fetch(API_PATH + '/newsletter/' + email, { method: 'POST' });
+    return this.fetch(API_PATH + '/newsletter/' + email, { method: 'POST' })
   }
 
   saveAvatar(type: 'default' | 'file' | 'gravatar', file?: Blob) {
@@ -278,11 +289,11 @@ export default class API {
             body: file,
           }
         : {}),
-    }).then(body => JSON.parse(body));
+    }).then(body => JSON.parse(body))
   }
 
   getJob(jobId: number) {
-    return this.fetch(`${API_PATH}/job/${jobId}`);
+    return this.fetch(`${API_PATH}/job/${jobId}`)
   }
 
   saveAvatarClip(blob: Blob): Promise<void> {
@@ -294,34 +305,34 @@ export default class API {
       body: blob,
     })
       .then(body => body)
-      .catch(err => err);
+      .catch(err => err)
   }
 
   fetchAvatarClip() {
-    return this.fetch(API_PATH + '/user_client/avatar_clip');
+    return this.fetch(API_PATH + '/user_client/avatar_clip')
   }
 
   deleteAvatarClip() {
-    return this.fetch(API_PATH + '/user_client/delete_avatar_clip');
+    return this.fetch(API_PATH + '/user_client/delete_avatar_clip')
   }
 
   fetchTakeouts() {
-    return this.fetch(API_PATH + '/user_client/takeout');
+    return this.fetch(API_PATH + '/user_client/takeout')
   }
 
   requestTakeout() {
     return this.fetch(API_PATH + '/user_client/takeout/request', {
       method: 'POST',
-    });
+    })
   }
 
-  fetchTakeoutLinks(id: number) {
+  fetchTakeoutLinks(id: number): Promise<TakeoutResponse> {
     return this.fetch(
       [API_PATH, 'user_client', 'takeout', id, 'links'].join('/'),
       {
         method: 'POST',
       }
-    );
+    )
   }
 
   fetchLeaderboard(type: 'clip' | 'vote', cursor?: [number, number]) {
@@ -330,27 +341,27 @@ export default class API {
         (type == 'clip' ? '' : '/votes') +
         '/leaderboard' +
         (cursor ? '?cursor=' + JSON.stringify(cursor) : '')
-    );
+    )
   }
 
   createGoal(locale: string, body: CustomGoalParams): Promise<void> {
     return this.fetch([API_PATH, 'user_client', locale, 'goals'].join('/'), {
       method: 'POST',
       body,
-    });
+    })
   }
 
   fetchGoals(locale?: string): Promise<AllGoals> {
     return this.fetch(
       API_PATH + '/user_client' + (locale ? '/' + locale : '') + '/goals'
-    );
+    )
   }
 
   claimAccount(): Promise<void> {
     return this.fetch(
       API_PATH + '/user_clients/' + this.user.userId + '/claim',
       { method: 'POST' }
-    );
+    )
   }
 
   saveHasDownloaded(
@@ -361,7 +372,7 @@ export default class API {
     return this.fetch(this.getLocalePath() + '/downloaders', {
       method: 'POST',
       body: { email, locale, dataset },
-    });
+    })
   }
 
   seenAwards(kind: 'award' | 'notification' = 'award'): Promise<void> {
@@ -370,7 +381,7 @@ export default class API {
         '/user_client/awards/seen' +
         (kind == 'notification' ? '?notification' : ''),
       { method: 'POST' }
-    );
+    )
   }
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -378,29 +389,29 @@ export default class API {
     return this.fetch(API_PATH + '/reports', {
       method: 'POST',
       body,
-    });
+    })
   }
 
   // Challenge
   fetchChallengePoints(): Promise<{
-    user: number;
-    team: number;
+    user: number
+    team: number
   }> {
     if (getChallenge(this.user)) {
       return this.fetch(
         `${API_PATH}/challenge/${this.user.account.enrollment.challenge}/points`
-      );
+      )
     }
-    return null;
+    return null
   }
 
   fetchWeeklyProgress(): Promise<WeeklyChallenge> {
     if (getChallenge(this.user)) {
       return this.fetch(
         `${API_PATH}/challenge/${this.user.account.enrollment.challenge}/progress`
-      );
+      )
     }
-    return null;
+    return null
   }
 
   fetchTopTeams(
@@ -412,9 +423,9 @@ export default class API {
         `${API_PATH}/challenge/${
           this.user.account.enrollment.challenge
         }/${locale}/teams?cursor=${cursor ? JSON.stringify(cursor) : ''}`
-      );
+      )
     }
-    return null;
+    return null
   }
 
   fetchTopContributors(
@@ -429,9 +440,9 @@ export default class API {
         }/${locale}/contributors/${type}?cursor=${
           cursor ? JSON.stringify(cursor) : ''
         }`
-      );
+      )
     }
-    return null;
+    return null
   }
 
   fetchTopMembers(
@@ -446,22 +457,22 @@ export default class API {
         }/${locale}/members/${type}?cursor=${
           cursor ? JSON.stringify(cursor) : ''
         }`
-      );
+      )
     }
-    return null;
+    return null
   }
   // check whether or not is the first invite
   fetchInviteStatus(): Promise<{
-    showInviteSendToast: boolean;
-    hasEarnedSessionToast: boolean;
-    challengeEnded: boolean;
+    showInviteSendToast: boolean
+    hasEarnedSessionToast: boolean
+    challengeEnded: boolean
   }> {
     if (getChallenge(this.user)) {
       return this.fetch(
         `${API_PATH}/challenge/${this.user.account.enrollment.challenge}/achievement/invite`
-      );
+      )
     }
-    return null;
+    return null
   }
 
   // Tell back-end user get unexpected achievement: invite + contribute in the same session
@@ -473,49 +484,49 @@ export default class API {
         {
           method: 'POST',
         }
-      );
+      )
     }
-    return null;
+    return null
   }
 
   getPublicUrl(path: string, bucketType: string): Promise<{ url: string }> {
     return this.fetch(`${API_PATH}/bucket/${bucketType}/${path}`, {
       method: 'GET',
-    });
+    })
   }
 
   async getServerDate(): Promise<string> {
-    return await this.fetch(`${API_PATH}/server_date`);
+    return await this.fetch(`${API_PATH}/server_date`)
   }
 
   getAccents(lang?: string) {
-    return this.fetch(`${API_PATH}/language/accents${lang ? '/' + lang : ''}`);
+    return this.fetch(`${API_PATH}/language/accents${lang ? '/' + lang : ''}`)
   }
 
   getVariants(lang?: string) {
-    return this.fetch(`${API_PATH}/language/variants${lang ? '/' + lang : ''}`);
+    return this.fetch(`${API_PATH}/language/variants${lang ? '/' + lang : ''}`)
   }
 
   getDatasets(releaseType: string) {
-    const query = releaseType ? `?releaseType=${releaseType}` : '';
-    return this.fetch(`${API_PATH}/datasets${query ? query : ''}`);
+    const query = releaseType ? `?releaseType=${releaseType}` : ''
+    return this.fetch(`${API_PATH}/datasets${query ? query : ''}`)
   }
 
   getLanguagesWithDatasets() {
-    return this.fetch(`${API_PATH}/datasets/languages`);
+    return this.fetch(`${API_PATH}/datasets/languages`)
   }
 
   getLanguageDatasetStats(languageCode: string) {
-    return this.fetch(`${API_PATH}/datasets/languages/${languageCode}`);
+    return this.fetch(`${API_PATH}/datasets/languages/${languageCode}`)
   }
   async sendLanguageRequest({
     email,
     languageInfo,
     languageLocale,
   }: {
-    email: string;
-    languageInfo: string;
-    languageLocale: string;
+    email: string
+    languageInfo: string
+    languageLocale: string
   }) {
     return this.fetch(`${API_PATH}/language/request`, {
       method: 'POST',
@@ -524,6 +535,60 @@ export default class API {
         languageInfo,
         languageLocale,
       },
-    });
+    })
+  }
+
+  createSentence({
+    sentence,
+    source,
+    localeId,
+    localeName,
+    domain,
+  }: SentenceSubmission) {
+    const data = { domain, sentence, source, localeId, localeName }
+
+    return this.fetch(`${API_PATH}/sentences`, {
+      method: 'POST',
+      body: data,
+    })
+  }
+
+  fetchPendingSentences(localeId: number) {
+    return this.fetch(`${API_PATH}/sentences/review?localeId=${localeId}`)
+  }
+
+  voteSentence({ vote, sentence_id }: SentenceVote) {
+    const data = { vote, sentence_id }
+
+    return this.fetch(`${API_PATH}/sentences/vote`, {
+      method: 'POST',
+      body: data,
+    })
+  }
+
+  async bulkSubmissionRequest({
+    file,
+    locale,
+    fileName,
+  }: {
+    file: File
+    locale: string
+    fileName: string
+  }) {
+    const { signal } = this.abortController
+    const content = await file.arrayBuffer()
+
+    return fetch(`${API_PATH}/${locale}/bulk_submissions`, {
+      method: 'POST',
+      body: content,
+      headers: {
+        filename: fileName,
+      },
+      signal,
+    })
+  }
+
+  abortBulkSubmissionRequest() {
+    this.abortController.abort()
   }
 }
